Create comparator sort coordinates.java
Comparator Sort Coordinates

You are given an array arr[] of n coordinates. These coordinates are objects of the class Point and contain x and y. You need to sort the array by increasing order of x, and if two x are same then sort by increasing order of y.

Examples:

Input: arr[] = [[1, 2], [3, 4], [5, 6], [7, 8], [9, 0]]
Output: (1,2) (3,4) (5,6) (7,8) (9,0)
Explanation: Sort it in increasing order of x, and if two x are same then sort in increasing order of y. So, by following the logic we get (1,2) (3,4) (5,6) (7,8) (9,0)