Create binary search .java
Binary Search Method

Given a sorted integer array arr[] and an integer element k. The task is to check if the element is present in the array or not and return the index.

Example 1:

Input: arr[] = [1, 2, 3, 4, 6], k = 6
Output: 4
Explanation: Since, 6 is present in the array at index 4 (0-based indexing), so output is 4.