Create equals method.java
Equals Method

Given a two integer arrays a[] and b[], you need to find if both the arrays contain the same elements in the same order or not.

Example 1:

Input: a[] = [1, 2, 3, 4], b[] = [4, 3, 2, 1]
Output: false
Explanation: Both arrays are not same as they don't contain elements in the same order. So we print false.